+++ MagneticField/Portable/test/BuildFile.xml
@@ -1,9 +1,9 @@
-<bin file="alpaka/testParabolicParametrizedMagneticField.dev.cc">
+<bin file="alpaka/testParabolicParametrizedTrackerField.dev.cc">
   <use name="alpaka"/>
   <use name="eigen"/>
   <use name="DataFormats/GeometryVector"/>
   <use name="FWCore/Utilities"/>
   <use name="HeterogeneousCore/AlpakaInterface"/>
-  <use name="MagneticField/ParametrizedEngine" source_only="1"/>
+  <use name="MagneticField/Portable" source_only="1"/>
   <flags ALPAKA_BACKENDS="1"/>
 </bin>
